Fixed ReadAll
* Initial DB files added

Added the dependencies and created an initial file for the database.

* Basic DB Implementation

Added the basic features of a database:
insert, read, read all, update and delete. It also has basic functionality to open and close the database.

* Update transaction.dart

Fixed an error where the fromJson method would read the double as an int

* Update transaction.dart

Fully fixed the bug where the db couldn't read account double values, with some try catch statments. Hopefully this is the end of this issue

* Increases functionality of 'readAll()'

- by default sorts by recenency
- can be limited to x most recent weeks
- can be limited to just one account

* Fix readAll

Attempt at a bug fix where the query to read data in from the query wasn't working correctly.

* Added Validator

Added a validator, basically confirms that an input is formatted correctly before it is added to the database

Co-authored-by: William McCormick <wmccormick1000@gmail.com>